vmbackup: moved snapshot to lib directory
