+++ lib/snapshot/snapshot_test.go
